backend/requirements.txt
langchain-google-genai==0.0.4
google-generativeai==0.3.1
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.14
fastapi==0.110.0
uvicorn==0.29.0
google-api-python-client==2.126.0
google-auth==2.29.0
langchain==0.1.14
langchain-google-genai==0.0.4
google-generativeai==0.3.1
streamlit==1.34.0
python-dotenv
langchain==0.1.14
langchain-google-genai==0.0.4
google-generativeai==0.3.1
streamlit==1.34.0
python-dotenv
